moved alarm
