+++ app/src/main/java/com/example/eyedroptracker/fragments/HomeFragment.kt
@@ -32,22 +32,6 @@ class HomeFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        /*
-        // This is to show to set an alarm.
-        var alarmService = AlarmService(requireContext())
-        fab.setOnClickListener{
-            setAlarm {
-                alarmService.setExactAlarm(it, "title", "message", 1)
-            }
-        }
-
-        // This is to show how you can delete an alarm.
-        fab2.setOnClickListener{
-            setAlarm {
-                alarmService.cancelExactAlarm(it, "title", "message", 1)
-            }
-        }
-        */
 
 
 //        save_user.setOnClickListener {
@@ -82,39 +66,6 @@ class HomeFragment : Fragment() {
 //        }
     }
 
-    /*
-    // This was taken from the youtube tutorial referenced in the README file.
-    private fun setAlarm(callback: (Long) -> Unit) {
-        Calendar.getInstance().apply {
-            this.set(Calendar.SECOND, 0)
-            this.set(Calendar.MILLISECOND, 0)
-            DatePickerDialog(
-                requireContext(),
-                0,
-                { _, year, month, day ->
-                    this.set(Calendar.YEAR, year)
-                    this.set(Calendar.MONTH, month)
-                    this.set(Calendar.DAY_OF_MONTH, day)
-                    TimePickerDialog(
-                        requireContext(),
-                        0,
-                        { _, hour, minute ->
-                            this.set(Calendar.HOUR_OF_DAY, hour)
-                            this.set(Calendar.MINUTE, minute)
-                            callback(this.timeInMillis)
-                        },
-                        this.get(Calendar.HOUR_OF_DAY),
-                        this.get(Calendar.MINUTE),
-                        false
-                    ).show()
-                },
-                this.get(Calendar.YEAR),
-                this.get(Calendar.MONTH),
-                this.get(Calendar.DAY_OF_MONTH)
-            ).show()
-        }
-    }
-     */
 
 
 }
