+++ app/src/main/java/com/example/eyedroptracker/fragments/ReminderFragment.kt
@@ -1,11 +1,20 @@
 package com.example.eyedroptracker.fragments
 
+import android.app.DatePickerDialog
+import android.app.TimePickerDialog
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.eyedroptracker.R
+import com.example.eyedroptracker.models.User
+import com.example.eyedroptracker.service.AlarmService
+import com.example.eyedroptracker.utils.AppDatabase
+import kotlinx.android.synthetic.main.fragment_home.*
+import kotlinx.android.synthetic.main.fragment_reminder.*
+import kotlinx.coroutines.runBlocking
+import java.util.*
 
 class ReminderFragment : Fragment() {
 
@@ -16,4 +25,59 @@ class ReminderFragment : Fragment() {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_reminder, container, false)
     }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+
+        // This is to show to set an alarm.
+        val alarmService = AlarmService(requireContext())
+        fab.setOnClickListener{
+            setAlarm {
+                alarmService.setExactAlarm(it, "title", "message", 1)
+            }
+        }
+
+        // This is to show how you can delete an alarm.
+        fab2.setOnClickListener{
+            setAlarm {
+                alarmService.cancelExactAlarm(it, "title", "message", 1)
+            }
+        }
+
+
+
+
+    }
+    private fun setAlarm(callback: (Long) -> Unit) {
+        Calendar.getInstance().apply {
+            this.set(Calendar.SECOND, 0)
+            this.set(Calendar.MILLISECOND, 0)
+            DatePickerDialog(
+                requireContext(),
+                0,
+                { _, year, month, day ->
+                    this.set(Calendar.YEAR, year)
+                    this.set(Calendar.MONTH, month)
+                    this.set(Calendar.DAY_OF_MONTH, day)
+                    TimePickerDialog(
+                        requireContext(),
+                        0,
+                        { _, hour, minute ->
+                            this.set(Calendar.HOUR_OF_DAY, hour)
+                            this.set(Calendar.MINUTE, minute)
+                            callback(this.timeInMillis)
+                        },
+                        this.get(Calendar.HOUR_OF_DAY),
+                        this.get(Calendar.MINUTE),
+                        false
+                    ).show()
+                },
+                this.get(Calendar.YEAR),
+                this.get(Calendar.MONTH),
+                this.get(Calendar.DAY_OF_MONTH)
+            ).show()
+        }
+    }
+
 }
